requirements.txt
absl==0.0
imgaug==0.4.0
keras==2.4.3
matplotlib==3.3.4
numpy==1.20.1
PyYAML==5.4.1
scikit_learn==0.24.1
scipy==1.6.1
tensorflow==2.4.1
tqdm==4.59.0
